@@ -1,32 +1,35 @@
 name: flutter_quill
 description: A rich text editor supporting mobile and web (Demo App @ bulletjournal.us)
-version: 2.1.0
+version: 4.2.3
 #author: bulletjournal
 homepage: https://bulletjournal.us/home/index.html
 repository: https://github.com/singerdmx/flutter-quill
 
 environment:
   sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.17.0"
+  flutter: ">=2.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
   collection: ^1.15.0
-  flutter_colorpicker: ^0.6.0
-  flutter_keyboard_visibility: ^5.0.0
-  image_picker: ^0.8.2
+  flutter_colorpicker: ^1.0.3
+  flutter_keyboard_visibility: ^5.2.0
+  image_picker: ^0.8.5
   photo_view: ^0.13.0
-  quiver: ^3.0.0
+  quiver: ^3.1.0
   string_validator: ^0.3.0
   tuple: ^2.0.0
-  url_launcher: ^6.0.2
-  pedantic: ^1.11.0
-  video_player: ^2.1.10
-  characters: ^1.1.0
+  url_launcher: ^6.1.0
+  pedantic: ^1.11.1
+  video_player: ^2.4.0
+  characters: ^1.2.0
   youtube_player_flutter: ^8.0.0
   diff_match_patch: ^0.4.1
-  i18n_extension: ^4.1.3
+  i18n_extension: ^4.2.1
+  gallery_saver: ^2.3.2
+  device_info_plus: ^3.2.3
+  platform: ^3.1.0
 
 dev_dependencies:
   flutter_test:
