@@ -8,7 +8,6 @@ import '../../translations/toolbar.i18n.dart';
 import '../../utils/color.dart';
 import '../controller.dart';
 import '../toolbar.dart';
-import 'quill_icon_button.dart';
 
 /// Controls color styles.
 ///
@@ -38,7 +37,7 @@ class _ColorButtonState extends State<ColorButton> {
   late bool _isToggledColor;
   late bool _isToggledBackground;
   late bool _isWhite;
-  late bool _isWhitebackground;
+  late bool _isWhiteBackground;
 
   Style get _selectionStyle => widget.controller.getSelectionStyle();
 
@@ -50,7 +49,7 @@ class _ColorButtonState extends State<ColorButton> {
           widget.controller.getSelectionStyle().attributes);
       _isWhite = _isToggledColor &&
           _selectionStyle.attributes['color']!.value == '#ffffff';
-      _isWhitebackground = _isToggledBackground &&
+      _isWhiteBackground = _isToggledBackground &&
           _selectionStyle.attributes['background']!.value == '#ffffff';
     });
   }
@@ -62,7 +61,7 @@ class _ColorButtonState extends State<ColorButton> {
     _isToggledBackground = _getIsToggledBackground(_selectionStyle.attributes);
     _isWhite = _isToggledColor &&
         _selectionStyle.attributes['color']!.value == '#ffffff';
-    _isWhitebackground = _isToggledBackground &&
+    _isWhiteBackground = _isToggledBackground &&
         _selectionStyle.attributes['background']!.value == '#ffffff';
     widget.controller.addListener(_didChangeEditingValue);
   }
@@ -86,7 +85,7 @@ class _ColorButtonState extends State<ColorButton> {
           _getIsToggledBackground(_selectionStyle.attributes);
       _isWhite = _isToggledColor &&
           _selectionStyle.attributes['color']!.value == '#ffffff';
-      _isWhitebackground = _isToggledBackground &&
+      _isWhiteBackground = _isToggledBackground &&
           _selectionStyle.attributes['background']!.value == '#ffffff';
     }
   }
@@ -105,7 +104,7 @@ class _ColorButtonState extends State<ColorButton> {
         : (widget.iconTheme?.iconUnselectedColor ?? theme.iconTheme.color);
 
     final iconColorBackground =
-        _isToggledBackground && widget.background && !_isWhitebackground
+        _isToggledBackground && widget.background && !_isWhiteBackground
             ? stringToColor(_selectionStyle.attributes['background']!.value)
             : (widget.iconTheme?.iconUnselectedColor ?? theme.iconTheme.color);
 
@@ -113,7 +112,7 @@ class _ColorButtonState extends State<ColorButton> {
         ? stringToColor('#ffffff')
         : (widget.iconTheme?.iconUnselectedFillColor ?? theme.canvasColor);
     final fillColorBackground =
-        _isToggledBackground && widget.background && _isWhitebackground
+        _isToggledBackground && widget.background && _isWhiteBackground
             ? stringToColor('#ffffff')
             : (widget.iconTheme?.iconUnselectedFillColor ?? theme.canvasColor);
 
@@ -125,6 +124,7 @@ class _ColorButtonState extends State<ColorButton> {
           size: widget.iconSize,
           color: widget.background ? iconColorBackground : iconColor),
       fillColor: widget.background ? fillColorBackground : fillColor,
+      borderRadius: widget.iconTheme?.borderRadius ?? 2,
       onPressed: _showColorPicker,
     );
   }
