@@ -2,7 +2,7 @@ import 'dart:math' as math;
 
 import '../../quill_delta.dart';
 import '../style.dart';
-import 'embed.dart';
+import 'embeddable.dart';
 import 'line.dart';
 import 'node.dart';
 
@@ -101,6 +101,13 @@ abstract class Leaf extends Node {
     }
   }
 
+  @override
+  String toString() {
+    final keys = style.keys.toList(growable: false)..sort();
+    final styleKeys = keys.join();
+    return '⟨$value⟩$styleKeys';
+  }
+
   /// Adjust this text node by merging it with adjacent nodes if they share
   /// the same style.
   @override
@@ -151,12 +158,23 @@ abstract class Leaf extends Node {
       return isLast ? null : next as Leaf?;
     }
 
-    assert(this is Text);
-    final text = _value as String;
-    _value = text.substring(0, index);
-    final split = Leaf(text.substring(index))..applyStyle(style);
-    insertAfter(split);
-    return split;
+    // assert(this is Text);
+    // 修改，由于flutter-quill和tun-editor共用一套document model，
+    // tun-editor插入@和#插入的方式是MentionEmbed（lib/src/models/quill_delta.dart:78），导致这里assert(this is Text)跑错误
+    // 所以临时兼容MentionEmbed，让Text和MentionEmbed可以同时存在，假如弃用tun-editor则可以还原此处
+    if (this is Embed && _value is MentionEmbed) {
+      final text = (_value as MentionEmbed).value;
+      final split = Leaf(_value)..applyStyle(style);
+      insertAfter(split);
+      return split;
+    } else {
+      assert(this is Text);
+      final text = _value as String;
+      _value = text.substring(0, index);
+      final split = Leaf(text.substring(index))..applyStyle(style);
+      insertAfter(split);
+      return split;
+    }
   }
 
   /// Cuts a leaf from [index] to the end of this node and returns new node
@@ -241,7 +259,9 @@ class Text extends Leaf {
 class Embed extends Leaf {
   Embed(Embeddable data) : super.val(data);
 
+  // Refer to https://www.fileformat.info/info/unicode/char/fffc/index.htm
   static const kObjectReplacementCharacter = '\uFFFC';
+  static const kObjectReplacementInt = 65532;
 
   @override
   Node newInstance() => throw UnimplementedError();
@@ -249,17 +269,24 @@ class Embed extends Leaf {
   @override
   Embeddable get value => super.value as Embeddable;
 
-  /// // Embed nodes are represented as unicode object replacement character in
+  // 修改，toPlainText返回值，兼容@、#
+  // Embed nodes are represented as unicode object replacement character in
   // plain text.
   @override
-  String toPlainText() => kObjectReplacementCharacter;
+  String toPlainText() => toContent();
 
   // 修改，添加@和#的embed类型不显示文本
   @override
   String toContent() {
-    if(value is MentionEmbed) {
+    if (value is MentionEmbed) {
       return (value as MentionEmbed).value;
     }
     return kObjectReplacementCharacter;
   }
+
+  @override
+  String toString() => '${super.toString()} ${value.type}';
+
+  @override
+  int get length => value.length;
 }
