@@ -1,14 +1,16 @@
 import 'dart:math' as math;
 
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/services.dart';
 import 'package:tuple/tuple.dart';
 
 import '../models/documents/attribute.dart';
 import '../models/documents/document.dart';
-import '../models/documents/nodes/embed.dart';
+import '../models/documents/nodes/embeddable.dart';
+import '../models/documents/nodes/leaf.dart';
 import '../models/documents/style.dart';
 import '../models/quill_delta.dart';
-import '../utils/diff_delta.dart';
+import '../utils/delta.dart';
 
 typedef ReplaceTextCallback = bool Function(int index, int len, Object? data);
 typedef DeleteCallback = void Function(int cursorPosition, bool forward);
@@ -20,6 +22,8 @@ class QuillController extends ChangeNotifier {
     bool keepStyleOnNewLine = false,
     this.onReplaceText,
     this.onDelete,
+    this.onSelectionCompleted,
+    this.onSelectionChanged,
   })  : _selection = selection,
         _keepStyleOnNewLine = keepStyleOnNewLine;
 
@@ -47,6 +51,9 @@ class QuillController extends ChangeNotifier {
 
   /// Custom delete handler
   DeleteCallback? onDelete;
+
+  void Function()? onSelectionCompleted;
+  void Function(TextSelection textSelection)? onSelectionChanged;
 
   /// Store any styles attribute that got toggled by the tap of a button
   /// and that has not been applied yet.
@@ -79,6 +86,20 @@ class QuillController extends ChangeNotifier {
     return document
         .collectStyle(selection.start, selection.end - selection.start)
         .mergeAll(toggledStyle);
+  }
+
+  /// Returns all styles for each node within selection
+  List<Tuple2<int, Style>> getAllIndividualSelectionStyles() {
+    final styles = document.collectAllIndividualStyles(
+        selection.start, selection.end - selection.start);
+    return styles;
+  }
+
+  /// Returns plain text for each node within selection
+  String getPlainText() {
+    final text =
+        document.getPlainText(selection.start, selection.end - selection.start);
+    return text;
   }
 
   /// Returns all styles for any character within the specified text range.
@@ -121,6 +142,12 @@ class QuillController extends ChangeNotifier {
   bool get hasUndo => document.hasUndo;
 
   bool get hasRedo => document.hasRedo;
+
+  /// clear editor
+  void clear() {
+    replaceText(0, plainTextEditingValue.text.length - 1, '',
+        const TextSelection.collapsed(offset: 0));
+  }
 
   void replaceText(
       int index, int len, Object? data, TextSelection? textSelection,
@@ -194,17 +221,30 @@ class QuillController extends ChangeNotifier {
   /// Called in two cases:
   /// forward == false && textBefore.isEmpty
   /// forward == true && textAfter.isEmpty
+  /// Android only
+  /// see https://github.com/singerdmx/flutter-quill/discussions/514
   void handleDelete(int cursorPosition, bool forward) =>
       onDelete?.call(cursorPosition, forward);
+
+  void formatTextStyle(int index, int len, Style style) {
+    style.attributes.forEach((key, attr) {
+      formatText(index, len, attr);
+    });
+  }
 
   void formatText(int index, int len, Attribute? attribute) {
     if (len == 0 &&
         attribute!.isInline &&
         attribute.key != Attribute.link.key) {
+      // Add the attribute to our toggledStyle.
+      // It will be used later upon insertion.
       toggledStyle = toggledStyle.put(attribute);
     }
 
     final change = document.format(index, len, attribute);
+    // Transform selection against the composed change and give priority to
+    // the change. This is needed in cases when format operation actually
+    // inserts data into the document (e.g. embeds).
     final adjustedSelection = selection.copyWith(
         baseOffset: change.transformPosition(selection.baseOffset),
         extentOffset: change.transformPosition(selection.extentOffset));
@@ -216,6 +256,22 @@ class QuillController extends ChangeNotifier {
 
   void formatSelection(Attribute? attribute) {
     formatText(selection.start, selection.end - selection.start, attribute);
+  }
+
+  void moveCursorToStart() {
+    updateSelection(
+        const TextSelection.collapsed(offset: 0), ChangeSource.LOCAL);
+  }
+
+  void moveCursorToPosition(int position) {
+    updateSelection(
+        TextSelection.collapsed(offset: position), ChangeSource.LOCAL);
+  }
+
+  void moveCursorToEnd() {
+    updateSelection(
+        TextSelection.collapsed(offset: plainTextEditingValue.text.length),
+        ChangeSource.LOCAL);
   }
 
   void updateSelection(TextSelection textSelection, ChangeSource source) {
@@ -273,5 +329,26 @@ class QuillController extends ChangeNotifier {
     _selection = selection.copyWith(
         baseOffset: math.min(selection.baseOffset, end),
         extentOffset: math.min(selection.extentOffset, end));
+    toggledStyle = Style();
+    onSelectionChanged?.call(textSelection);
   }
+
+  /// Given offset, find its leaf node in document
+  Leaf? queryNode(int offset) {
+    return document.querySegmentLeafNode(offset).item2;
+  }
+
+  /// Clipboard for image url and its corresponding style
+  /// item1 is url and item2 is style string
+  Tuple2<String, String>? _copiedImageUrl;
+
+  Tuple2<String, String>? get copiedImageUrl => _copiedImageUrl;
+
+  set copiedImageUrl(Tuple2<String, String>? value) {
+    _copiedImageUrl = value;
+    Clipboard.setData(const ClipboardData(text: ''));
+  }
+
+  // Notify toolbar buttons directly with attributes
+  Map<String, Attribute> toolbarButtonToggler = {};
 }
