@@ -1,5 +1,3 @@
-import 'dart:convert';
-import 'dart:io' as io;
 import 'dart:math' as math;
 
 import 'package:flutter/cupertino.dart';
@@ -8,64 +6,49 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
-import 'package:string_validator/string_validator.dart';
-import 'package:url_launcher/url_launcher.dart';
+import 'package:i18n_extension/i18n_widget.dart';
+import 'package:tuple/tuple.dart';
 
-import '../models/documents/attribute.dart';
 import '../models/documents/document.dart';
 import '../models/documents/nodes/container.dart' as container_node;
-import '../models/documents/nodes/embed.dart';
-import '../models/documents/nodes/leaf.dart' as leaf;
-import '../models/documents/nodes/line.dart';
-import '../utils/string_helper.dart';
+import '../models/documents/style.dart';
+import '../utils/platform.dart';
 import 'box.dart';
 import 'controller.dart';
 import 'cursor.dart';
 import 'default_styles.dart';
 import 'delegate.dart';
-import 'image.dart';
+import 'embeds/default_embed_builder.dart';
+import 'float_cursor.dart';
+import 'link.dart';
 import 'raw_editor.dart';
 import 'text_selection.dart';
-import 'video_app.dart';
-import 'youtube_video_app.dart';
 
-const linkPrefixes = [
-  'mailto:', // email
-  'tel:', // telephone
-  'sms:', // SMS
-  'callto:',
-  'wtai:',
-  'market:',
-  'geopoint:',
-  'ymsgr:',
-  'msnim:',
-  'gtalk:', // Google Talk
-  'skype:',
-  'sip:', // Lync
-  'whatsapp:',
-  'http'
-];
-
-abstract class EditorState extends State<RawEditor> {
+/// Base interface for the editor state which defines contract used by
+/// various mixins.
+abstract class EditorState extends State<RawEditor>
+    implements TextSelectionDelegate {
   ScrollController get scrollController;
 
-  TextEditingValue getTextEditingValue();
+  RenderEditor get renderEditor;
 
-  void setTextEditingValue(TextEditingValue value);
+  EditorTextSelectionOverlay? get selectionOverlay;
 
-  RenderEditor? getRenderEditor();
+  List<Tuple2<int, Style>> get pasteStyle;
 
-  EditorTextSelectionOverlay? getSelectionOverlay();
+  String get pastePlainText;
+
+  /// Controls the floating cursor animation when it is released.
+  /// The floating cursor is animated to merge with the regular cursor.
+  AnimationController get floatingCursorResetController;
 
   bool showToolbar();
-
-  void hideToolbar();
 
   void requestKeyboard();
 }
 
 /// Base interface for editable render objects.
-abstract class RenderAbstractEditor {
+abstract class RenderAbstractEditor implements TextLayoutMetrics {
   TextSelection selectWordAtPosition(TextPosition position);
 
   TextSelection selectLineAtPosition(TextPosition position);
@@ -87,8 +70,23 @@ abstract class RenderAbstractEditor {
   /// selection that contains some text but whose ends meet in the middle).
   TextPosition getPositionForOffset(Offset offset);
 
+  /// Returns the local coordinates of the endpoints of the given selection.
+  ///
+  /// If the selection is collapsed (and therefore occupies a single point), the
+  /// returned list is of length one. Otherwise, the selection is not collapsed
+  /// and the returned list is of length two. In this case, however, the two
+  /// points might actually be co-located (e.g., because of a bidirectional
+  /// selection that contains some text but whose ends meet in the middle).
   List<TextSelectionPoint> getEndpointsForSelection(
       TextSelection textSelection);
+
+  /// Sets the screen position of the floating cursor and the text position
+  /// closest to the cursor.
+  /// `resetLerpValue` drives the size of the floating cursor.
+  /// See [EditorState.floatingCursorResetController].
+  void setFloatingCursor(FloatingCursorDragState dragState,
+      Offset lastBoundedOffset, TextPosition lastTextPosition,
+      {double? resetLerpValue});
 
   /// If [ignorePointer] is false (the default) then this method is called by
   /// the internal gesture recognizer's [TapGestureRecognizer.onTapDown]
@@ -115,8 +113,13 @@ abstract class RenderAbstractEditor {
   /// {@macro flutter.rendering.editable.select}
   void selectWordEdge(SelectionChangedCause cause);
 
-  /// Select text between the global positions [from] and [to].
-  void selectPositionAt(Offset from, Offset to, SelectionChangedCause cause);
+  ///
+  /// Returns the new selection. Note that the returned value may not be
+  /// yet reflected in the latest widget state.
+  ///
+  /// Returns null if no change occurred.
+  TextSelection? selectPositionAt(
+      {required Offset from, required SelectionChangedCause cause, Offset? to});
 
   /// Select a word around the location of the last tap down.
   ///
@@ -133,95 +136,7 @@ abstract class RenderAbstractEditor {
   /// If you have a [TextEditingController], it's generally easier to
   /// programmatically manipulate its `value` or `selection` directly.
   /// {@endtemplate}
-  void selectPosition(SelectionChangedCause cause);
-}
-
-String _standardizeImageUrl(String url) {
-  if (url.contains('base64')) {
-    return url.split(',')[1];
-  }
-  return url;
-}
-
-bool _isMobile() => io.Platform.isAndroid || io.Platform.isIOS;
-
-Widget _defaultEmbedBuilder(
-    BuildContext context, leaf.Embed node, bool readOnly) {
-  assert(!kIsWeb, 'Please provide EmbedBuilder for Web');
-  switch (node.value.type) {
-    case 'image':
-      final imageUrl = _standardizeImageUrl(node.value.data);
-
-      final style = node.style.attributes['style'];
-      if (_isMobile() && style != null) {
-        final _attrs = parseKeyValuePairs(style.value.toString(),
-            {'mobileWidth', 'mobileHeight', 'mobileMargin', 'mobileAlignment'});
-        if (_attrs.isNotEmpty) {
-          assert(
-              _attrs['mobileWidth'] != null && _attrs['mobileHeight'] != null,
-              'mobileWidth and mobileHeight must be specified');
-          final w = double.parse(_attrs['mobileWidth']!);
-          final h = double.parse(_attrs['mobileHeight']!);
-          final m = _attrs['mobileMargin'] == null
-              ? 0.0
-              : double.parse(_attrs['mobileMargin']!);
-          var a = Alignment.center;
-          if (_attrs['mobileAlignment'] != null) {
-            final _index = [
-              'topLeft',
-              'topCenter',
-              'topRight',
-              'centerLeft',
-              'center',
-              'centerRight',
-              'bottomLeft',
-              'bottomCenter',
-              'bottomRight'
-            ].indexOf(_attrs['mobileAlignment']!);
-            if (_index >= 0) {
-              a = [
-                Alignment.topLeft,
-                Alignment.topCenter,
-                Alignment.topRight,
-                Alignment.centerLeft,
-                Alignment.center,
-                Alignment.centerRight,
-                Alignment.bottomLeft,
-                Alignment.bottomCenter,
-                Alignment.bottomRight
-              ][_index];
-            }
-          }
-          return Padding(
-              padding: EdgeInsets.all(m),
-              child: imageUrl.startsWith('http')
-                  ? Image.network(imageUrl, width: w, height: h, alignment: a)
-                  : isBase64(imageUrl)
-                      ? Image.memory(base64.decode(imageUrl),
-                          width: w, height: h, alignment: a)
-                      : Image.file(io.File(imageUrl),
-                          width: w, height: h, alignment: a));
-        }
-      }
-      return imageUrl.startsWith('http')
-          ? Image.network(imageUrl)
-          : isBase64(imageUrl)
-              ? Image.memory(base64.decode(imageUrl))
-              : Image.file(io.File(imageUrl));
-    case 'video':
-      final videoUrl = node.value.data;
-      if (videoUrl.contains('youtube.com') || videoUrl.contains('youtu.be')) {
-        return YoutubeVideoApp(
-            videoUrl: videoUrl, context: context, readOnly: readOnly);
-      }
-      return VideoApp(videoUrl: videoUrl, context: context, readOnly: readOnly);
-    default:
-      throw UnimplementedError(
-        'Embeddable type "${node.value.type}" is not supported by default '
-        'embed builder of QuillEditor. You must pass your own builder function '
-        'to embedBuilder property of QuillEditor or QuillField widgets.',
-      );
-  }
+  void selectPosition({required SelectionChangedCause cause});
 }
 
 class QuillEditor extends StatefulWidget {
@@ -234,6 +149,8 @@ class QuillEditor extends StatefulWidget {
       required this.autoFocus,
       required this.readOnly,
       required this.expands,
+      this.isSelectionInViewport,
+      this.caretOffset = 0,
       this.showCursor,
       this.paintCursorAboveText,
       this.placeholder,
@@ -241,6 +158,7 @@ class QuillEditor extends StatefulWidget {
       this.scrollBottomInset = 0,
       this.minHeight,
       this.maxHeight,
+      this.maxContentWidth,
       this.customStyles,
       this.textCapitalization = TextCapitalization.sentences,
       this.keyboardAppearance = Brightness.light,
@@ -251,14 +169,20 @@ class QuillEditor extends StatefulWidget {
       this.onSingleLongTapStart,
       this.onSingleLongTapMoveUpdate,
       this.onSingleLongTapEnd,
-      this.embedBuilder = _defaultEmbedBuilder,
+      this.embedBuilder = defaultEmbedBuilder,
+      this.linkActionPickerDelegate = defaultLinkActionPickerDelegate,
       this.customStyleBuilder,
-      Key? key});
+      this.locale,
+      this.floatingCursorDisabled = false,
+      this.textSelectionControls,
+      Key? key})
+      : super(key: key);
 
   factory QuillEditor.basic({
     required QuillController controller,
     required bool readOnly,
     Brightness? keyboardAppearance,
+    IsSelectionInViewport? isSelectionInViewport,
   }) {
     return QuillEditor(
       controller: controller,
@@ -270,29 +194,134 @@ class QuillEditor extends StatefulWidget {
       expands: false,
       padding: EdgeInsets.zero,
       keyboardAppearance: keyboardAppearance ?? Brightness.light,
+      isSelectionInViewport: isSelectionInViewport,
     );
   }
 
+  /// Controller object which establishes a link between a rich text document
+  /// and this editor.
+  ///
+  /// Must not be null.
   final QuillController controller;
+
+  /// Controls whether this editor has keyboard focus.
   final FocusNode focusNode;
+
+  /// The [ScrollController] to use when vertically scrolling the contents.
   final ScrollController scrollController;
+
+  /// Whether this editor should create a scrollable container for its content.
+  ///
+  /// When set to `true` the editor's height can be controlled by [minHeight],
+  /// [maxHeight] and [expands] properties.
+  ///
+  /// When set to `false` the editor always expands to fit the entire content
+  /// of the document and should normally be placed as a child of another
+  /// scrollable widget, otherwise the content may be clipped.
   final bool scrollable;
   final double scrollBottomInset;
+
+  /// Additional space around the content of this editor.
   final EdgeInsetsGeometry padding;
+
+  /// Whether this editor should focus itself if nothing else is already
+  /// focused.
+  ///
+  /// If true, the keyboard will open as soon as this editor obtains focus.
+  /// Otherwise, the keyboard is only shown after the user taps the editor.
+  ///
+  /// Defaults to `false`. Cannot be `null`.
   final bool autoFocus;
+
+  /// Whether to show cursor.
+  ///
+  /// The cursor refers to the blinking caret when the editor is focused.
   final bool? showCursor;
   final bool? paintCursorAboveText;
+
+  /// Whether the text can be changed.
+  ///
+  /// When this is set to `true`, the text cannot be modified
+  /// by any shortcut or keyboard operation. The text is still selectable.
+  ///
+  /// Defaults to `false`. Must not be `null`.
   final bool readOnly;
   final String? placeholder;
+
+  /// Whether to enable user interface affordances for changing the
+  /// text selection.
+  ///
+  /// For example, setting this to true will enable features such as
+  /// long-pressing the editor to select text and show the
+  /// cut/copy/paste menu, and tapping to move the text cursor.
+  ///
+  /// When this is false, the text selection cannot be adjusted by
+  /// the user, text cannot be copied, and the user cannot paste into
+  /// the text field from the clipboard.
   final bool enableInteractiveSelection;
+
+  /// The minimum height to be occupied by this editor.
+  ///
+  /// This only has effect if [scrollable] is set to `true` and [expands] is
+  /// set to `false`.
   final double? minHeight;
+
+  /// The maximum height to be occupied by this editor.
+  ///
+  /// This only has effect if [scrollable] is set to `true` and [expands] is
+  /// set to `false`.
   final double? maxHeight;
+
+  /// The maximum width to be occupied by the content of this editor.
+  ///
+  /// If this is not null and and this editor's width is larger than this value
+  /// then the contents will be constrained to the provided maximum width and
+  /// horizontally centered. This is mostly useful on devices with wide screens.
+  final double? maxContentWidth;
+
   final DefaultStyles? customStyles;
+
+  /// Whether this editor's height will be sized to fill its parent.
+  ///
+  /// This only has effect if [scrollable] is set to `true`.
+  ///
+  /// If expands is set to true and wrapped in a parent widget like [Expanded]
+  /// or [SizedBox], the editor will expand to fill the parent.
+  ///
+  /// [maxHeight] and [minHeight] must both be `null` when this is set to
+  /// `true`.
+  ///
+  /// Defaults to `false`.
   final bool expands;
+
+  /// Configures how the platform keyboard will select an uppercase or
+  /// lowercase keyboard.
+  ///
+  /// Only supports text keyboards, other keyboard types will ignore this
+  /// configuration. Capitalization is locale-aware.
+  ///
+  /// Defaults to [TextCapitalization.sentences]. Must not be `null`.
   final TextCapitalization textCapitalization;
+
+  /// The appearance of the keyboard.
+  ///
+  /// This setting is only honored on iOS devices.
+  ///
+  /// Defaults to [Brightness.light].
   final Brightness keyboardAppearance;
+
+  /// The [ScrollPhysics] to use when vertically scrolling the input.
+  ///
+  /// This only has effect if [scrollable] is set to `true`.
+  ///
+  /// If not specified, it will behave according to the current platform.
+  ///
+  /// See [Scrollable.physics].
   final ScrollPhysics? scrollPhysics;
+
+  /// Callback to invoke when user wants to launch a URL.
   final ValueChanged<String>? onLaunchUrl;
+
   // Returns whether gesture is handled
   final bool Function(
       TapDownDetails details, TextPosition Function(Offset offset))? onTapDown;
@@ -317,11 +346,41 @@ class QuillEditor extends StatefulWidget {
   final EmbedBuilder embedBuilder;
   final CustomStyleBuilder? customStyleBuilder;
 
+  /// The locale to use for the editor toolbar, defaults to system locale
+  /// More https://github.com/singerdmx/flutter-quill#translation
+  final Locale? locale;
+
+  /// Delegate function responsible for showing menu with link actions on
+  /// mobile platforms (iOS, Android).
+  ///
+  /// The menu is triggered in editing mode ([readOnly] is set to `false`)
+  /// when the user long-presses a link-styled text segment.
+  ///
+  /// FlutterQuill provides default implementation which can be overridden by
+  /// this field to customize the user experience.
+  ///
+  /// By default on iOS the menu is displayed with [showCupertinoModalPopup]
+  /// which constructs an instance of [CupertinoActionSheet]. For Android,
+  /// the menu is displayed with [showModalBottomSheet] and a list of
+  /// Material [ListTile]s.
+  final LinkActionPickerDelegate linkActionPickerDelegate;
+
+  final bool floatingCursorDisabled;
+
+  final IsSelectionInViewport? isSelectionInViewport;
+  
+  final double caretOffset; 
+
+  /// allows to create a custom textSelectionControls,
+  /// if this is null a default textSelectionControls based on the app's theme
+  /// will be used
+  final TextSelectionControls? textSelectionControls;
+
   @override
-  _QuillEditorState createState() => _QuillEditorState();
+  QuillEditorState createState() => QuillEditorState();
 }
 
-class _QuillEditorState extends State<QuillEditor>
+class QuillEditorState extends State<QuillEditor>
     implements EditorTextSelectionGestureDetectorBuilderDelegate {
   final GlobalKey<EditorState> _editorKey = GlobalKey<EditorState>();
   late EditorTextSelectionGestureDetectorBuilder
@@ -347,98 +406,106 @@ class _QuillEditorState extends State<QuillEditor>
     Color selectionColor;
     Radius? cursorRadius;
 
-    switch (theme.platform) {
-      case TargetPlatform.android:
-      case TargetPlatform.fuchsia:
-      case TargetPlatform.linux:
-      case TargetPlatform.windows:
-        textSelectionControls = materialTextSelectionControls;
-        paintCursorAboveText = false;
-        cursorOpacityAnimates = false;
-        cursorColor ??= selectionTheme.cursorColor ?? theme.colorScheme.primary;
-        selectionColor = selectionTheme.selectionColor ??
-            theme.colorScheme.primary.withOpacity(0.40);
-        break;
-      case TargetPlatform.iOS:
-      case TargetPlatform.macOS:
-        final cupertinoTheme = CupertinoTheme.of(context);
-        textSelectionControls = cupertinoTextSelectionControls;
-        paintCursorAboveText = true;
-        cursorOpacityAnimates = true;
-        cursorColor ??=
-            selectionTheme.cursorColor ?? cupertinoTheme.primaryColor;
-        selectionColor = selectionTheme.selectionColor ??
-            cupertinoTheme.primaryColor.withOpacity(0.40);
-        cursorRadius ??= const Radius.circular(2);
-        cursorOffset = Offset(
-            iOSHorizontalOffset / MediaQuery.of(context).devicePixelRatio, 0);
-        break;
-      default:
-        throw UnimplementedError();
+    if (isAppleOS(theme.platform)) {
+      final cupertinoTheme = CupertinoTheme.of(context);
+      textSelectionControls = cupertinoTextSelectionControls;
+      paintCursorAboveText = true;
+      cursorOpacityAnimates = true;
+      cursorColor ??= selectionTheme.cursorColor ?? cupertinoTheme.primaryColor;
+      selectionColor = selectionTheme.selectionColor ??
+          cupertinoTheme.primaryColor.withOpacity(0.40);
+      cursorRadius ??= const Radius.circular(2);
+      cursorOffset = Offset(
+          iOSHorizontalOffset / MediaQuery.of(context).devicePixelRatio, 0);
+    } else {
+      textSelectionControls = materialTextSelectionControls;
+      paintCursorAboveText = false;
+      cursorOpacityAnimates = false;
+      cursorColor ??= selectionTheme.cursorColor ?? theme.colorScheme.primary;
+      selectionColor = selectionTheme.selectionColor ??
+          theme.colorScheme.primary.withOpacity(0.40);
     }
 
-    return _selectionGestureDetectorBuilder.build(
-      HitTestBehavior.translucent,
-      RawEditor(
-        _editorKey,
-        widget.controller,
-        widget.focusNode,
-        widget.scrollController,
-        widget.scrollable,
-        widget.scrollBottomInset,
-        widget.padding,
-        widget.readOnly,
-        widget.placeholder,
-        widget.onLaunchUrl,
-        ToolbarOptions(
-          copy: widget.enableInteractiveSelection,
-          cut: widget.enableInteractiveSelection,
-          paste: widget.enableInteractiveSelection,
-          selectAll: widget.enableInteractiveSelection,
-        ),
-        theme.platform == TargetPlatform.iOS ||
-            theme.platform == TargetPlatform.android,
-        widget.showCursor,
-        CursorStyle(
-          color: cursorColor,
-          backgroundColor: Colors.grey,
-          width: 2,
-          radius: cursorRadius,
-          offset: cursorOffset,
-          paintAboveText: widget.paintCursorAboveText ?? paintCursorAboveText,
-          opacityAnimates: cursorOpacityAnimates,
-        ),
-        widget.textCapitalization,
-        widget.maxHeight,
-        widget.minHeight,
-        widget.customStyles,
-        widget.expands,
-        widget.autoFocus,
-        selectionColor,
-        textSelectionControls,
-        widget.keyboardAppearance,
-        widget.enableInteractiveSelection,
-        widget.scrollPhysics,
-        widget.embedBuilder,
-        widget.customStyleBuilder,
+    final child = RawEditor(
+      key: _editorKey,
+      controller: widget.controller,
+      focusNode: widget.focusNode,
+      scrollController: widget.scrollController,
+      scrollable: widget.scrollable,
+      scrollBottomInset: widget.scrollBottomInset,
+      padding: widget.padding,
+      readOnly: widget.readOnly,
+      placeholder: widget.placeholder,
+      onLaunchUrl: widget.onLaunchUrl,
+      toolbarOptions: ToolbarOptions(
+        copy: widget.enableInteractiveSelection,
+        cut: widget.enableInteractiveSelection,
+        paste: widget.enableInteractiveSelection,
+        selectAll: widget.enableInteractiveSelection,
       ),
+      showSelectionHandles: isMobile(theme.platform),
+      showCursor: widget.showCursor,
+      cursorStyle: CursorStyle(
+        color: cursorColor,
+        backgroundColor: Colors.grey,
+        width: 2,
+        radius: cursorRadius,
+        offset: cursorOffset,
+        paintAboveText: widget.paintCursorAboveText ?? paintCursorAboveText,
+        opacityAnimates: cursorOpacityAnimates,
+      ),
+      textCapitalization: widget.textCapitalization,
+      minHeight: widget.minHeight,
+      maxHeight: widget.maxHeight,
+      maxContentWidth: widget.maxContentWidth,
+      customStyles: widget.customStyles,
+      expands: widget.expands,
+      autoFocus: widget.autoFocus,
+      selectionColor: selectionColor,
+      selectionCtrls: widget.textSelectionControls ?? textSelectionControls,
+      keyboardAppearance: widget.keyboardAppearance,
+      enableInteractiveSelection: widget.enableInteractiveSelection,
+      scrollPhysics: widget.scrollPhysics,
+      embedBuilder: widget.embedBuilder,
+      linkActionPickerDelegate: widget.linkActionPickerDelegate,
+      customStyleBuilder: widget.customStyleBuilder,
+      floatingCursorDisabled: widget.floatingCursorDisabled,
+      isSelectionInViewport: widget.isSelectionInViewport,
+      caretOffset: widget.caretOffset,
     );
+
+    final editor = I18n(
+        initialLocale: widget.locale,
+        child: _selectionGestureDetectorBuilder.build(
+          behavior: HitTestBehavior.translucent,
+          child: child,
+        ));
+
+    if (kIsWeb) {
+      // Intercept RawKeyEvent on Web to prevent it from propagating to parents
+      // that might interfere with the editor key behavior, such as
+      // SingleChildScrollView. Thanks to @wliumelb for the workaround.
+      // See issue https://github.com/singerdmx/flutter-quill/issues/304
+      return RawKeyboardListener(
+        onKey: (_) {},
+        focusNode: FocusNode(
+          onKey: (node, event) => KeyEventResult.skipRemainingHandlers,
+        ),
+        child: editor,
+      );
+    }
+
+    return editor;
   }
 
   @override
-  GlobalKey<EditorState> getEditableTextKey() {
-    return _editorKey;
-  }
+  GlobalKey<EditorState> get editableTextKey => _editorKey;
 
   @override
-  bool getForcePressEnabled() {
-    return false;
-  }
+  bool get forcePressEnabled => false;
 
   @override
-  bool getSelectionEnabled() {
-    return widget.enableInteractiveSelection;
-  }
+  bool get selectionEnabled => widget.enableInteractiveSelection;
 
   void _requestKeyboard() {
     _editorKey.currentState!.requestKeyboard();
@@ -447,15 +514,16 @@ class _QuillEditorState extends State<QuillEditor>
 
 class _QuillEditorSelectionGestureDetectorBuilder
     extends EditorTextSelectionGestureDetectorBuilder {
-  _QuillEditorSelectionGestureDetectorBuilder(this._state) : super(_state);
+  _QuillEditorSelectionGestureDetectorBuilder(this._state)
+      : super(delegate: _state);
 
-  final _QuillEditorState _state;
+  final QuillEditorState _state;
 
   @override
   void onForcePressStart(ForcePressDetails details) {
     super.onForcePressStart(details);
-    if (delegate.getSelectionEnabled() && shouldShowSelectionToolbar) {
-      getEditor()!.showToolbar();
+    if (delegate.selectionEnabled && shouldShowSelectionToolbar) {
+      editor!.showToolbar();
     }
   }
 
@@ -465,191 +533,143 @@ class _QuillEditorSelectionGestureDetectorBuilder
   @override
   void onSingleLongTapMoveUpdate(LongPressMoveUpdateDetails details) {
     if (_state.widget.onSingleLongTapMoveUpdate != null) {
-      final renderEditor = getRenderEditor();
-      if (renderEditor != null) {
-        if (_state.widget.onSingleLongTapMoveUpdate!(
-            details, renderEditor.getPositionForOffset)) {
-          return;
-        }
+      if (renderEditor != null &&
+          _state.widget.onSingleLongTapMoveUpdate!(
+              details, renderEditor!.getPositionForOffset)) {
+        return;
       }
     }
-    if (!delegate.getSelectionEnabled()) {
+    if (!delegate.selectionEnabled) {
       return;
     }
-    switch (Theme.of(_state.context).platform) {
-      case TargetPlatform.iOS:
-      case TargetPlatform.macOS:
-        getRenderEditor()!.selectPositionAt(
-          details.globalPosition,
-          null,
-          SelectionChangedCause.longPress,
-        );
-        break;
-      case TargetPlatform.android:
-      case TargetPlatform.fuchsia:
-      case TargetPlatform.linux:
-      case TargetPlatform.windows:
-        getRenderEditor()!.selectWordsInRange(
-          details.globalPosition - details.offsetFromOrigin,
-          details.globalPosition,
-          SelectionChangedCause.longPress,
-        );
-        break;
-      default:
-        throw 'Invalid platform';
+
+    final _platform = Theme.of(_state.context).platform;
+    if (isAppleOS(_platform)) {
+      renderEditor!.selectPositionAt(
+        from: details.globalPosition,
+        cause: SelectionChangedCause.longPress,
+      );
+    } else {
+      renderEditor!.selectWordsInRange(
+        details.globalPosition - details.offsetFromOrigin,
+        details.globalPosition,
+        SelectionChangedCause.longPress,
+      );
     }
   }
 
-  bool _onTapping(TapUpDetails details) {
+  bool _isPositionSelected(TapUpDetails details) {
     if (_state.widget.controller.document.isEmpty()) {
       return false;
     }
-    final pos = getRenderEditor()!.getPositionForOffset(details.globalPosition);
+    final pos = renderEditor!.getPositionForOffset(details.globalPosition);
     final result =
-        getEditor()!.widget.controller.document.queryChild(pos.offset);
-    if (result.node == null) {
+        editor!.widget.controller.document.querySegmentLeafNode(pos.offset);
+    final line = result.item1;
+    if (line == null) {
       return false;
     }
-    final line = result.node as Line;
-    final segmentResult = line.queryChild(result.offset, false);
-    if (segmentResult.node == null) {
-      if (line.length == 1) {
-        getEditor()!.widget.controller.updateSelection(
-            TextSelection.collapsed(offset: pos.offset), ChangeSource.LOCAL);
-        return true;
-      }
-      return false;
+    final segmentLeaf = result.item2;
+    if (segmentLeaf == null && line.length == 1) {
+      editor!.widget.controller.updateSelection(
+          TextSelection.collapsed(offset: pos.offset), ChangeSource.LOCAL);
+      return true;
     }
-    final segment = segmentResult.node as leaf.Leaf;
-    if (segment.style.containsKey(Attribute.link.key)) {
-      var launchUrl = getEditor()!.widget.onLaunchUrl;
-      launchUrl ??= _launchUrl;
-      String? link = segment.style.attributes[Attribute.link.key]!.value;
-      if (getEditor()!.widget.readOnly && link != null) {
-        link = link.trim();
-        if (!linkPrefixes
-            .any((linkPrefix) => link!.toLowerCase().startsWith(linkPrefix))) {
-          link = 'https://$link';
-        }
-        launchUrl(link);
-      }
-      return false;
-    }
-    if (getEditor()!.widget.readOnly && segment.value is BlockEmbed) {
-      final blockEmbed = segment.value as BlockEmbed;
-      if (blockEmbed.type == 'image') {
-        final imageUrl = _standardizeImageUrl(blockEmbed.data);
-        Navigator.push(
-          getEditor()!.context,
-          MaterialPageRoute(
-            builder: (context) => ImageTapWrapper(
-              imageProvider: imageUrl.startsWith('http')
-                  ? NetworkImage(imageUrl)
-                  : isBase64(imageUrl)
-                      ? Image.memory(base64.decode(imageUrl))
-                          as ImageProvider<Object>?
-                      : FileImage(io.File(imageUrl)),
-            ),
-          ),
-        );
-      }
-    }
-
     return false;
-  }
-
-  Future<void> _launchUrl(String url) async {
-    await launch(url);
   }
 
   @override
   void onTapDown(TapDownDetails details) {
     if (_state.widget.onTapDown != null) {
-      final renderEditor = getRenderEditor();
-      if (renderEditor != null) {
-        if (_state.widget.onTapDown!(
-            details, renderEditor.getPositionForOffset)) {
-          return;
-        }
+      if (renderEditor != null &&
+          _state.widget.onTapDown!(
+              details, renderEditor!.getPositionForOffset)) {
+        return;
       }
     }
     super.onTapDown(details);
   }
 
+  bool isShiftClick(PointerDeviceKind deviceKind) {
+    final pressed = RawKeyboard.instance.keysPressed;
+    return deviceKind == PointerDeviceKind.mouse &&
+        (pressed.contains(LogicalKeyboardKey.shiftLeft) ||
+            pressed.contains(LogicalKeyboardKey.shiftRight));
+  }
+
   @override
   void onSingleTapUp(TapUpDetails details) {
-    if (_state.widget.onTapUp != null) {
-      final renderEditor = getRenderEditor();
-      if (renderEditor != null) {
-        if (_state.widget.onTapUp!(
-            details, renderEditor.getPositionForOffset)) {
-          return;
-        }
-      }
+    if (_state.widget.onTapUp != null &&
+        renderEditor != null &&
+        _state.widget.onTapUp!(details, renderEditor!.getPositionForOffset)) {
+      return;
     }
 
-    getEditor()!.hideToolbar();
+    editor!.hideToolbar();
 
-    final positionSelected = _onTapping(details);
-
-    if (delegate.getSelectionEnabled() && !positionSelected) {
-      switch (Theme.of(_state.context).platform) {
-        case TargetPlatform.iOS:
-        case TargetPlatform.macOS:
+    try {
+      if (delegate.selectionEnabled && !_isPositionSelected(details)) {
+        final _platform = Theme.of(_state.context).platform;
+        if (isAppleOS(_platform)) {
           switch (details.kind) {
             case PointerDeviceKind.mouse:
             case PointerDeviceKind.stylus:
             case PointerDeviceKind.invertedStylus:
-              getRenderEditor()!.selectPosition(SelectionChangedCause.tap);
+              // Precise devices should place the cursor at a precise position.
+              // If `Shift` key is pressed then
+              // extend current selection instead.
+              if (isShiftClick(details.kind)) {
+                renderEditor!
+                  ..extendSelection(details.globalPosition,
+                      cause: SelectionChangedCause.tap)
+                  ..onSelectionCompleted();
+              } else {
+                renderEditor!
+                  ..selectPosition(cause: SelectionChangedCause.tap)
+                  ..onSelectionCompleted();
+              }
+
               break;
             case PointerDeviceKind.touch:
             case PointerDeviceKind.unknown:
-              getRenderEditor()!.selectWordEdge(SelectionChangedCause.tap);
+              // On macOS/iOS/iPadOS a touch tap places the cursor at the edge
+              // of the word.
+              renderEditor!
+                ..selectWordEdge(SelectionChangedCause.tap)
+                ..onSelectionCompleted();
               break;
           }
-          break;
-        case TargetPlatform.android:
-        case TargetPlatform.fuchsia:
-        case TargetPlatform.linux:
-        case TargetPlatform.windows:
-          getRenderEditor()!.selectPosition(SelectionChangedCause.tap);
-          break;
+        } else {
+          renderEditor!
+            ..selectPosition(cause: SelectionChangedCause.tap)
+            ..onSelectionCompleted();
+        }
       }
+    } finally {
+      _state._requestKeyboard();
     }
-    _state._requestKeyboard();
   }
 
   @override
   void onSingleLongTapStart(LongPressStartDetails details) {
     if (_state.widget.onSingleLongTapStart != null) {
-      final renderEditor = getRenderEditor();
-      if (renderEditor != null) {
-        if (_state.widget.onSingleLongTapStart!(
-            details, renderEditor.getPositionForOffset)) {
-          return;
-        }
+      if (renderEditor != null &&
+          _state.widget.onSingleLongTapStart!(
+              details, renderEditor!.getPositionForOffset)) {
+        return;
       }
     }
 
-    if (delegate.getSelectionEnabled()) {
-      switch (Theme.of(_state.context).platform) {
-        case TargetPlatform.iOS:
-        case TargetPlatform.macOS:
-          getRenderEditor()!.selectPositionAt(
-            details.globalPosition,
-            null,
-            SelectionChangedCause.longPress,
-          );
-          break;
-        case TargetPlatform.android:
-        case TargetPlatform.fuchsia:
-        case TargetPlatform.linux:
-        case TargetPlatform.windows:
-          getRenderEditor()!.selectWord(SelectionChangedCause.longPress);
-          Feedback.forLongPress(_state.context);
-          break;
-        default:
-          throw 'Invalid platform';
+    if (delegate.selectionEnabled) {
+      final _platform = Theme.of(_state.context).platform;
+      if (isAppleOS(_platform)) {
+        renderEditor!.selectPositionAt(
+          from: details.globalPosition,
+          cause: SelectionChangedCause.longPress,
+        );
+      } else {
+        renderEditor!.selectWord(SelectionChangedCause.longPress);
+        Feedback.forLongPress(_state.context);
       }
     }
   }
@@ -657,11 +677,14 @@ class _QuillEditorSelectionGestureDetectorBuilder
   @override
   void onSingleLongTapEnd(LongPressEndDetails details) {
     if (_state.widget.onSingleLongTapEnd != null) {
-      final renderEditor = getRenderEditor();
       if (renderEditor != null) {
         if (_state.widget.onSingleLongTapEnd!(
-            details, renderEditor.getPositionForOffset)) {
+            details, renderEditor!.getPositionForOffset)) {
           return;
+        }
+
+        if (delegate.selectionEnabled) {
+          renderEditor!.onSelectionCompleted();
         }
       }
     }
@@ -669,38 +692,92 @@ class _QuillEditorSelectionGestureDetectorBuilder
   }
 }
 
+/// Signature for the callback that reports when the user changes the selection
+/// (including the cursor location).
+///
+/// Used by [RenderEditor.onSelectionChanged].
 typedef TextSelectionChangedHandler = void Function(
     TextSelection selection, SelectionChangedCause cause);
 
+/// Signature for the callback that reports when a selection action is actually
+/// completed and ratified. Completion is defined as when the user input has
+/// concluded for an entire selection action. For simple taps and keyboard input
+/// events that change the selection, this callback is invoked immediately
+/// following the TextSelectionChangedHandler. For long taps, the selection is
+/// considered complete at the up event of a long tap. For drag selections, the
+/// selection completes once the drag/pan event ends or is interrupted.
+///
+/// Used by [RenderEditor.onSelectionCompleted].
+typedef TextSelectionCompletedHandler = void Function();
+
+// The padding applied to text field. Used to determine the bounds when
+// moving the floating cursor.
+const EdgeInsets _kFloatingCursorAddedMargin = EdgeInsets.fromLTRB(4, 4, 4, 5);
+
+// The additional size on the x and y axis with which to expand the prototype
+// cursor to render the floating cursor in pixels.
+const EdgeInsets _kFloatingCaretSizeIncrease =
+    EdgeInsets.symmetric(horizontal: 0.5, vertical: 1);
+
+typedef IsSelectionInViewport = List<bool> Function(double selectionStart, double selectionEnd);
+
+/// Displays a document as a vertical list of document segments (lines
+/// and blocks).
+///
+/// Children of [RenderEditor] must be instances of [RenderEditableBox].
 class RenderEditor extends RenderEditableContainerBox
+    with RelayoutWhenSystemFontsChangeMixin
     implements RenderAbstractEditor {
-  RenderEditor(
+  RenderEditor({
+    required this.key,
+    required this.document,
+    required TextDirection textDirection,
+    required bool hasFocus,
+    required this.selection,
+    required this.scrollable,
+    required LayerLink startHandleLayerLink,
+    required LayerLink endHandleLayerLink,
+    required EdgeInsetsGeometry padding,
+    required CursorCont cursorController,
+    required this.onSelectionChanged,
+    required this.onSelectionCompleted,
+    required double scrollBottomInset,
+    required this.floatingCursorDisabled,
+    this.isSelectionInViewport,
     ViewportOffset? offset,
     List<RenderEditableBox>? children,
-    TextDirection textDirection,
-    double scrollBottomInset,
-    EdgeInsetsGeometry padding,
-    this.document,
-    this.selection,
-    this._hasFocus,
-    this.onSelectionChanged,
-    this._startHandleLayerLink,
-    this._endHandleLayerLink,
-    EdgeInsets floatingCursorAddedMargin,
-  ) : super(
-          children,
-          document.root,
-          textDirection,
-          scrollBottomInset,
-          padding,
+    EdgeInsets floatingCursorAddedMargin =
+        const EdgeInsets.fromLTRB(4, 4, 4, 5),
+    double? maxContentWidth,
+  })  : _hasFocus = hasFocus,
+        _extendSelectionOrigin = selection,
+        _startHandleLayerLink = startHandleLayerLink,
+        _endHandleLayerLink = endHandleLayerLink,
+        _cursorController = cursorController,
+        _maxContentWidth = maxContentWidth,
+        super(
+          children: children,
+          container: document.root,
+          textDirection: textDirection,
+          scrollBottomInset: scrollBottomInset,
+          padding: padding,
         );
 
+  final CursorCont _cursorController;
+  final bool floatingCursorDisabled;
+  final bool scrollable;
+
   Document document;
+  GlobalKey key;
   TextSelection selection;
   bool _hasFocus = false;
   LayerLink _startHandleLayerLink;
   LayerLink _endHandleLayerLink;
+  IsSelectionInViewport? isSelectionInViewport;
+
+  /// Called when the selection changes.
   TextSelectionChangedHandler onSelectionChanged;
+  TextSelectionCompletedHandler onSelectionCompleted;
   final ValueNotifier<bool> _selectionStartInViewport =
       ValueNotifier<bool>(true);
 
@@ -711,28 +788,49 @@ class RenderEditor extends RenderEditableContainerBox
   final ValueNotifier<bool> _selectionEndInViewport = ValueNotifier<bool>(true);
 
   void _updateSelectionExtentsVisibility(Offset effectiveOffset) {
-    final visibleRegion = Offset.zero & size;
-    final startPosition =
-        TextPosition(offset: selection.start, affinity: selection.affinity);
-    final startOffset = _getOffsetForCaret(startPosition);
-    // TODO(justinmc): https://github.com/flutter/flutter/issues/31495
-    // Check if the selection is visible with an approximation because a
-    // difference between rounded and unrounded values causes the caret to be
-    // reported as having a slightly (< 0.5) negative y offset. This rounding
-    // happens in paragraph.cc's layout and TextPainer's
-    // _applyFloatingPointHack. Ideally, the rounding mismatch will be fixed and
-    // this can be changed to be a strict check instead of an approximation.
-    const visibleRegionSlop = 0.5;
-    _selectionStartInViewport.value = visibleRegion
-        .inflate(visibleRegionSlop)
-        .contains(startOffset + effectiveOffset);
+    // 修改，添加isSelectionInViewport!=null的逻辑
+    if(isSelectionInViewport!=null) {
+      final renderBox = (key.currentContext!.findRenderObject()!) as RenderBox ;
+      final edotirOffset = renderBox.localToGlobal(Offset.zero);
 
-    final endPosition =
-        TextPosition(offset: selection.end, affinity: selection.affinity);
-    final endOffset = _getOffsetForCaret(endPosition);
-    _selectionEndInViewport.value = visibleRegion
-        .inflate(visibleRegionSlop)
-        .contains(endOffset + effectiveOffset);
+      final startPosition =
+      TextPosition(offset: selection.start, affinity: selection.affinity);
+      final startOffset = _getOffsetForCaret(startPosition);
+      final startPositionY = edotirOffset.dy + startOffset.dy;
+
+      final endPosition =
+      TextPosition(offset: selection.end, affinity: selection.affinity);
+      final endOffset = _getOffsetForCaret(endPosition);
+      final endPositionY = edotirOffset.dy + endOffset.dy;
+
+      final List<bool> res = isSelectionInViewport!.call(startPositionY,endPositionY);
+      _selectionStartInViewport.value = res[0];
+      _selectionEndInViewport.value = res[1];
+
+    } else {
+      final visibleRegion = Offset.zero & size;
+      final startPosition =
+      TextPosition(offset: selection.start, affinity: selection.affinity);
+      final startOffset = _getOffsetForCaret(startPosition);
+      // TODO(justinmc): https://github.com/flutter/flutter/issues/31495
+      // Check if the selection is visible with an approximation because a
+      // difference between rounded and unrounded values causes the caret to be
+      // reported as having a slightly (< 0.5) negative y offset. This rounding
+      // happens in paragraph.cc's layout and TextPainer's
+      // _applyFloatingPointHack. Ideally, the rounding mismatch will be fixed and
+      // this can be changed to be a strict check instead of an approximation.
+      const visibleRegionSlop = 0.5;
+      _selectionStartInViewport.value = visibleRegion
+          .inflate(visibleRegionSlop)
+          .contains(startOffset + effectiveOffset);
+
+      final endPosition =
+      TextPosition(offset: selection.end, affinity: selection.affinity);
+      final endOffset = _getOffsetForCaret(endPosition);
+      _selectionEndInViewport.value = visibleRegion
+          .inflate(visibleRegionSlop)
+          .contains(endOffset + effectiveOffset);
+    }
   }
 
   // returns offset relative to this at which the caret will be painted
@@ -780,7 +878,17 @@ class RenderEditor extends RenderEditableContainerBox
     }
     selection = t;
     markNeedsPaint();
+
+    if (!_shiftPressed && !_isDragging) {
+      // Only update extend selection origin if Shift key is not pressed and
+      // user is not dragging selection.
+      _extendSelectionOrigin = selection;
+    }
   }
+
+  bool get _shiftPressed =>
+      RawKeyboard.instance.keysPressed.contains(LogicalKeyboardKey.shiftLeft) ||
+      RawKeyboard.instance.keysPressed.contains(LogicalKeyboardKey.shiftRight);
 
   void setStartHandleLayerLink(LayerLink value) {
     if (_startHandleLayerLink == value) {
@@ -806,13 +914,20 @@ class RenderEditor extends RenderEditableContainerBox
     markNeedsPaint();
   }
 
+  double? _maxContentWidth;
+  set maxContentWidth(double? value) {
+    if (_maxContentWidth == value) return;
+    _maxContentWidth = value;
+    markNeedsLayout();
+  }
+
   @override
   List<TextSelectionPoint> getEndpointsForSelection(
       TextSelection textSelection) {
     if (textSelection.isCollapsed) {
       final child = childAtPosition(textSelection.extent);
       final localPosition = TextPosition(
-          offset: textSelection.extentOffset - child.getContainer().offset);
+          offset: textSelection.extentOffset - child.container.offset);
       final localOffset = child.getOffsetForCaret(localPosition);
       final parentData = child.parentData as BoxParentData;
       return <TextSelectionPoint>[
@@ -828,7 +943,7 @@ class RenderEditor extends RenderEditableContainerBox
 
     var baseChild = firstChild;
     while (baseChild != null) {
-      if (baseChild.getContainer() == baseNode) {
+      if (baseChild.container == baseNode) {
         break;
       }
       baseChild = childAfter(baseChild);
@@ -837,7 +952,7 @@ class RenderEditor extends RenderEditableContainerBox
 
     final baseParentData = baseChild!.parentData as BoxParentData;
     final baseSelection =
-        localSelection(baseChild.getContainer(), textSelection, true);
+        localSelection(baseChild.container, textSelection, true);
     var basePoint = baseChild.getBaseEndpointForSelection(baseSelection);
     basePoint = TextSelectionPoint(
         basePoint.point + baseParentData.offset, basePoint.direction);
@@ -845,7 +960,7 @@ class RenderEditor extends RenderEditableContainerBox
     final extentNode = _container.queryChild(textSelection.end, false).node;
     RenderEditableBox? extentChild = baseChild;
     while (extentChild != null) {
-      if (extentChild.getContainer() == extentNode) {
+      if (extentChild.container == extentNode) {
         break;
       }
       extentChild = childAfter(extentChild);
@@ -854,7 +969,7 @@ class RenderEditor extends RenderEditableContainerBox
 
     final extentParentData = extentChild!.parentData as BoxParentData;
     final extentSelection =
-        localSelection(extentChild.getContainer(), textSelection, true);
+        localSelection(extentChild.container, textSelection, true);
     var extentPoint =
         extentChild.getExtentEndpointForSelection(extentSelection);
     extentPoint = TextSelectionPoint(
@@ -865,9 +980,34 @@ class RenderEditor extends RenderEditableContainerBox
 
   Offset? _lastTapDownPosition;
 
+  // Used on Desktop (mouse and keyboard enabled platforms) as base offset
+  // for extending selection, either with combination of `Shift` + Click or
+  // by dragging
+  TextSelection? _extendSelectionOrigin;
+
   @override
   void handleTapDown(TapDownDetails details) {
     _lastTapDownPosition = details.globalPosition;
+  }
+
+  bool _isDragging = false;
+
+  void handleDragStart(DragStartDetails details) {
+    _isDragging = true;
+
+    final newSelection = selectPositionAt(
+      from: details.globalPosition,
+      cause: SelectionChangedCause.drag,
+    );
+
+    if (newSelection == null) return;
+    // Make sure to remember the origin for extend selection.
+    _extendSelectionOrigin = newSelection;
+  }
+
+  void handleDragEnd(DragEndDetails details) {
+    _isDragging = false;
+    onSelectionCompleted();
   }
 
   @override
@@ -906,12 +1046,40 @@ class RenderEditor extends RenderEditableContainerBox
     onSelectionChanged(nextSelection, cause);
   }
 
+  /// Extends current selection to the position closest to specified offset.
+  void extendSelection(Offset to, {required SelectionChangedCause cause}) {
+    /// The below logic does not exactly match the native version because
+    /// we do not allow swapping of base and extent positions.
+    assert(_extendSelectionOrigin != null);
+    final position = getPositionForOffset(to);
+
+    if (position.offset < _extendSelectionOrigin!.baseOffset) {
+      _handleSelectionChange(
+        TextSelection(
+          baseOffset: position.offset,
+          extentOffset: _extendSelectionOrigin!.extentOffset,
+          affinity: selection.affinity,
+        ),
+        cause,
+      );
+    } else if (position.offset > _extendSelectionOrigin!.extentOffset) {
+      _handleSelectionChange(
+        TextSelection(
+          baseOffset: _extendSelectionOrigin!.baseOffset,
+          extentOffset: position.offset,
+          affinity: selection.affinity,
+        ),
+        cause,
+      );
+    }
+  }
+
   @override
   void selectWordEdge(SelectionChangedCause cause) {
     assert(_lastTapDownPosition != null);
     final position = getPositionForOffset(_lastTapDownPosition!);
     final child = childAtPosition(position);
-    final nodeOffset = child.getContainer().offset;
+    final nodeOffset = child.container.offset;
     final localPosition = TextPosition(
       offset: position.offset - nodeOffset,
       affinity: position.affinity,
@@ -936,11 +1104,11 @@ class RenderEditor extends RenderEditableContainerBox
   }
 
   @override
-  void selectPositionAt(
-    Offset from,
+  TextSelection? selectPositionAt({
+    required Offset from,
+    required SelectionChangedCause cause,
     Offset? to,
-    SelectionChangedCause cause,
-  ) {
+  }) {
     final fromPosition = getPositionForOffset(from);
     final toPosition = to == null ? null : getPositionForOffset(to);
 
@@ -956,7 +1124,10 @@ class RenderEditor extends RenderEditableContainerBox
       extentOffset: extentOffset,
       affinity: fromPosition.affinity,
     );
+
+    // Call [onSelectionChanged] only when the selection actually changed.
     _handleSelectionChange(newSelection, cause);
+    return newSelection;
   }
 
   @override
@@ -965,21 +1136,14 @@ class RenderEditor extends RenderEditableContainerBox
   }
 
   @override
-  void selectPosition(SelectionChangedCause cause) {
-    selectPositionAt(_lastTapDownPosition!, null, cause);
+  void selectPosition({required SelectionChangedCause cause}) {
+    selectPositionAt(from: _lastTapDownPosition!, cause: cause);
   }
 
   @override
   TextSelection selectWordAtPosition(TextPosition position) {
-    final child = childAtPosition(position);
-    final nodeOffset = child.getContainer().offset;
-    final localPosition = TextPosition(
-        offset: position.offset - nodeOffset, affinity: position.affinity);
-    final localWord = child.getWordBoundary(localPosition);
-    final word = TextRange(
-      start: localWord.start + nodeOffset,
-      end: localWord.end + nodeOffset,
-    );
+    final word = getWordBoundary(position);
+    // When long-pressing past the end of the text, we want a collapsed cursor.
     if (position.offset >= word.end) {
       return TextSelection.fromPosition(position);
     }
@@ -988,16 +1152,9 @@ class RenderEditor extends RenderEditableContainerBox
 
   @override
   TextSelection selectLineAtPosition(TextPosition position) {
-    final child = childAtPosition(position);
-    final nodeOffset = child.getContainer().offset;
-    final localPosition = TextPosition(
-        offset: position.offset - nodeOffset, affinity: position.affinity);
-    final localLineRange = child.getLineBoundary(localPosition);
-    final line = TextRange(
-      start: localLineRange.start + nodeOffset,
-      end: localLineRange.end + nodeOffset,
-    );
+    final line = getLineAtOffset(position);
 
+    // When long-pressing past the end of the text, we want a collapsed cursor.
     if (position.offset >= line.end) {
       return TextSelection.fromPosition(position);
     }
@@ -1005,10 +1162,76 @@ class RenderEditor extends RenderEditableContainerBox
   }
 
   @override
+  void performLayout() {
+    assert(() {
+      if (!scrollable || !constraints.hasBoundedHeight) return true;
+      throw FlutterError.fromParts(<DiagnosticsNode>[
+        ErrorSummary('RenderEditableContainerBox must have '
+            'unlimited space along its main axis when it is scrollable.'),
+        ErrorDescription('RenderEditableContainerBox does not clip or'
+            ' resize its children, so it must be '
+            'placed in a parent that does not constrain the main '
+            'axis.'),
+        ErrorHint(
+            'You probably want to put the RenderEditableContainerBox inside a '
+            'RenderViewport with a matching main axis or disable the '
+            'scrollable property.')
+      ]);
+    }());
+    assert(() {
+      if (constraints.hasBoundedWidth) return true;
+      throw FlutterError.fromParts(<DiagnosticsNode>[
+        ErrorSummary('RenderEditableContainerBox must have a bounded'
+            ' constraint for its cross axis.'),
+        ErrorDescription('RenderEditableContainerBox forces its children to '
+            "expand to fit the RenderEditableContainerBox's container, "
+            'so it must be placed in a parent that constrains the cross '
+            'axis to a finite dimension.'),
+      ]);
+    }());
+
+    resolvePadding();
+    assert(resolvedPadding != null);
+
+    var mainAxisExtent = resolvedPadding!.top;
+    var child = firstChild;
+    final innerConstraints = BoxConstraints.tightFor(
+            width: math.min(
+                _maxContentWidth ?? double.infinity, constraints.maxWidth))
+        .deflate(resolvedPadding!);
+    final leftOffset = _maxContentWidth == null
+        ? 0.0
+        : math.max((constraints.maxWidth - _maxContentWidth!) / 2, 0);
+    while (child != null) {
+      child.layout(innerConstraints, parentUsesSize: true);
+      final childParentData = child.parentData as EditableContainerParentData
+        ..offset = Offset(resolvedPadding!.left + leftOffset, mainAxisExtent);
+      mainAxisExtent += child.size.height;
+      assert(child.parentData == childParentData);
+      child = childParentData.nextSibling;
+    }
+    mainAxisExtent += resolvedPadding!.bottom;
+    size = constraints.constrain(Size(constraints.maxWidth, mainAxisExtent));
+
+    assert(size.isFinite);
+  }
+
+  @override
   void paint(PaintingContext context, Offset offset) {
+    if (_hasFocus &&
+        _cursorController.show.value &&
+        !_cursorController.style.paintAboveText) {
+      _paintFloatingCursor(context, offset);
+    }
     defaultPaint(context, offset);
     _updateSelectionExtentsVisibility(offset + _paintOffset);
     _paintHandleLayers(context, getEndpointsForSelection(selection));
+
+    if (_hasFocus &&
+        _cursorController.show.value &&
+        _cursorController.style.paintAboveText) {
+      _paintFloatingCursor(context, offset);
+    }
   }
 
   @override
@@ -1046,19 +1269,19 @@ class RenderEditor extends RenderEditableContainerBox
   double preferredLineHeight(TextPosition position) {
     final child = childAtPosition(position);
     return child.preferredLineHeight(
-        TextPosition(offset: position.offset - child.getContainer().offset));
+        TextPosition(offset: position.offset - child.container.offset));
   }
 
   @override
   TextPosition getPositionForOffset(Offset offset) {
     final local = globalToLocal(offset);
-    final child = childAtOffset(local)!;
+    final child = childAtOffset(local);
 
     final parentData = child.parentData as BoxParentData;
     final localOffset = local - parentData.offset;
     final localPosition = child.getPositionForOffset(localOffset);
     return TextPosition(
-      offset: localPosition.offset + child.getContainer().offset,
+      offset: localPosition.offset + child.container.offset,
       affinity: localPosition.affinity,
     );
   }
@@ -1068,8 +1291,25 @@ class RenderEditor extends RenderEditableContainerBox
   /// The offset is the distance from the top of the editor and is the minimum
   /// from the current scroll position until [selection] becomes visible.
   /// Returns null if [selection] is already visible.
+  ///
+  /// Finds the closest scroll offset that fully reveals the editing cursor.
+  ///
+  /// The `scrollOffset` parameter represents current scroll offset in the
+  /// parent viewport.
+  ///
+  /// The `offsetInViewport` parameter represents the editor's vertical offset
+  /// in the parent viewport. This value should normally be 0.0 if this editor
+  /// is the only child of the viewport or if it's the topmost child. Otherwise
+  /// it should be a positive value equal to total height of all siblings of
+  /// this editor from above it.
+  ///
+  /// Returns `null` if the cursor is currently visible.
   double? getOffsetToRevealCursor(
       double viewportHeight, double scrollOffset, double offsetInViewport) {
+    // Endpoints coordinates represents lower left or lower right corner of
+    // the selection. If we want to scroll up to reveal the caret we need to
+    // adjust the dy value by the height of the line. We also add a small margin
+    // so that the caret is not too close to the edge of the viewport.
     final endpoints = getEndpointsForSelection(selection);
 
     // when we drag the right handle, we should get the last point
@@ -1086,13 +1326,13 @@ class RenderEditor extends RenderEditableContainerBox
       }
     }
 
+    // Collapsed selection => caret
     final child = childAtPosition(selection.extent);
     const kMargin = 8.0;
 
     final caretTop = endpoint.point.dy -
         child.preferredLineHeight(TextPosition(
-            offset:
-                selection.extentOffset - child.getContainer().documentOffset)) -
+            offset: selection.extentOffset - child.container.documentOffset)) -
         kMargin +
         offsetInViewport +
         scrollBottomInset;
@@ -1107,6 +1347,7 @@ class RenderEditor extends RenderEditableContainerBox
     if (dy == null) {
       return null;
     }
+    // Clamping to 0.0 so that the content does not jump unnecessarily.
     return math.max(dy, 0);
   }
 
@@ -1120,11 +1361,285 @@ class RenderEditor extends RenderEditableContainerBox
     final boxParentData = targetChild.parentData as BoxParentData;
     return childLocalRect.shift(Offset(0, boxParentData.offset.dy));
   }
+
+  // Start floating cursor
+
+  FloatingCursorPainter get _floatingCursorPainter => FloatingCursorPainter(
+        floatingCursorRect: _floatingCursorRect,
+        style: _cursorController.style,
+      );
+
+  bool _floatingCursorOn = false;
+  Rect? _floatingCursorRect;
+
+  TextPosition get floatingCursorTextPosition => _floatingCursorTextPosition;
+  late TextPosition _floatingCursorTextPosition;
+
+  // The relative origin in relation to the distance the user has theoretically
+  // dragged the floating cursor offscreen.
+  // This value is used to account for the difference
+  // in the rendering position and the raw offset value.
+  Offset _relativeOrigin = Offset.zero;
+  Offset? _previousOffset;
+  bool _resetOriginOnLeft = false;
+  bool _resetOriginOnRight = false;
+  bool _resetOriginOnTop = false;
+  bool _resetOriginOnBottom = false;
+
+  /// Returns the position within the editor closest to the raw cursor offset.
+  Offset calculateBoundedFloatingCursorOffset(
+      Offset rawCursorOffset, double preferredLineHeight) {
+    var deltaPosition = Offset.zero;
+    final topBound = _kFloatingCursorAddedMargin.top;
+    final bottomBound =
+        size.height - preferredLineHeight + _kFloatingCursorAddedMargin.bottom;
+    final leftBound = _kFloatingCursorAddedMargin.left;
+    final rightBound = size.width - _kFloatingCursorAddedMargin.right;
+
+    if (_previousOffset != null) {
+      deltaPosition = rawCursorOffset - _previousOffset!;
+    }
+
+    // If the raw cursor offset has gone off an edge,
+    // we want to reset the relative origin of
+    // the dragging when the user drags back into the field.
+    if (_resetOriginOnLeft && deltaPosition.dx > 0) {
+      _relativeOrigin =
+          Offset(rawCursorOffset.dx - leftBound, _relativeOrigin.dy);
+      _resetOriginOnLeft = false;
+    } else if (_resetOriginOnRight && deltaPosition.dx < 0) {
+      _relativeOrigin =
+          Offset(rawCursorOffset.dx - rightBound, _relativeOrigin.dy);
+      _resetOriginOnRight = false;
+    }
+    if (_resetOriginOnTop && deltaPosition.dy > 0) {
+      _relativeOrigin =
+          Offset(_relativeOrigin.dx, rawCursorOffset.dy - topBound);
+      _resetOriginOnTop = false;
+    } else if (_resetOriginOnBottom && deltaPosition.dy < 0) {
+      _relativeOrigin =
+          Offset(_relativeOrigin.dx, rawCursorOffset.dy - bottomBound);
+      _resetOriginOnBottom = false;
+    }
+
+    final currentX = rawCursorOffset.dx - _relativeOrigin.dx;
+    final currentY = rawCursorOffset.dy - _relativeOrigin.dy;
+    final double adjustedX =
+        math.min(math.max(currentX, leftBound), rightBound);
+    final double adjustedY =
+        math.min(math.max(currentY, topBound), bottomBound);
+    final adjustedOffset = Offset(adjustedX, adjustedY);
+
+    if (currentX < leftBound && deltaPosition.dx < 0) {
+      _resetOriginOnLeft = true;
+    } else if (currentX > rightBound && deltaPosition.dx > 0) {
+      _resetOriginOnRight = true;
+    }
+    if (currentY < topBound && deltaPosition.dy < 0) {
+      _resetOriginOnTop = true;
+    } else if (currentY > bottomBound && deltaPosition.dy > 0) {
+      _resetOriginOnBottom = true;
+    }
+
+    _previousOffset = rawCursorOffset;
+
+    return adjustedOffset;
+  }
+
+  @override
+  void setFloatingCursor(FloatingCursorDragState dragState,
+      Offset boundedOffset, TextPosition textPosition,
+      {double? resetLerpValue}) {
+    if (floatingCursorDisabled) return;
+
+    if (dragState == FloatingCursorDragState.Start) {
+      _relativeOrigin = Offset.zero;
+      _previousOffset = null;
+      _resetOriginOnBottom = false;
+      _resetOriginOnTop = false;
+      _resetOriginOnRight = false;
+      _resetOriginOnBottom = false;
+    }
+    _floatingCursorOn = dragState != FloatingCursorDragState.End;
+    if (_floatingCursorOn) {
+      _floatingCursorTextPosition = textPosition;
+      final sizeAdjustment = resetLerpValue != null
+          ? EdgeInsets.lerp(
+              _kFloatingCaretSizeIncrease, EdgeInsets.zero, resetLerpValue)!
+          : _kFloatingCaretSizeIncrease;
+      final child = childAtPosition(textPosition);
+      final caretPrototype =
+          child.getCaretPrototype(child.globalToLocalPosition(textPosition));
+      _floatingCursorRect =
+          sizeAdjustment.inflateRect(caretPrototype).shift(boundedOffset);
+      _cursorController
+          .setFloatingCursorTextPosition(_floatingCursorTextPosition);
+    } else {
+      _floatingCursorRect = null;
+      _cursorController.setFloatingCursorTextPosition(null);
+    }
+  }
+
+  void _paintFloatingCursor(PaintingContext context, Offset offset) {
+    _floatingCursorPainter.paint(context.canvas);
+  }
+
+  // End floating cursor
+
+  // Start TextLayoutMetrics implementation
+
+  /// Return a [TextSelection] containing the line of the given [TextPosition].
+  @override
+  TextSelection getLineAtOffset(TextPosition position) {
+    final child = childAtPosition(position);
+    final nodeOffset = child.container.offset;
+    final localPosition = TextPosition(
+        offset: position.offset - nodeOffset, affinity: position.affinity);
+    final localLineRange = child.getLineBoundary(localPosition);
+    final line = TextRange(
+      start: localLineRange.start + nodeOffset,
+      end: localLineRange.end + nodeOffset,
+    );
+    return TextSelection(baseOffset: line.start, extentOffset: line.end);
+  }
+
+  @override
+  TextRange getWordBoundary(TextPosition position) {
+    final child = childAtPosition(position);
+    final nodeOffset = child.container.offset;
+    final localPosition = TextPosition(
+        offset: position.offset - nodeOffset, affinity: position.affinity);
+    final localWord = child.getWordBoundary(localPosition);
+    return TextRange(
+      start: localWord.start + nodeOffset,
+      end: localWord.end + nodeOffset,
+    );
+  }
+
+  /// Returns the TextPosition above the given offset into the text.
+  ///
+  /// If the offset is already on the first line, the offset of the first
+  /// character will be returned.
+  @override
+  TextPosition getTextPositionAbove(TextPosition position) {
+    final child = childAtPosition(position);
+    final localPosition =
+        TextPosition(offset: position.offset - child.container.documentOffset);
+
+    var newPosition = child.getPositionAbove(localPosition);
+
+    if (newPosition == null) {
+      // There was no text above in the current child, check the direct
+      // sibling.
+      final sibling = childBefore(child);
+      if (sibling == null) {
+        // reached beginning of the document, move to the
+        // first character
+        newPosition = const TextPosition(offset: 0);
+      } else {
+        final caretOffset = child.getOffsetForCaret(localPosition);
+        final testPosition = TextPosition(offset: sibling.container.length - 1);
+        final testOffset = sibling.getOffsetForCaret(testPosition);
+        final finalOffset = Offset(caretOffset.dx, testOffset.dy);
+        final siblingPosition = sibling.getPositionForOffset(finalOffset);
+        newPosition = TextPosition(
+            offset: sibling.container.documentOffset + siblingPosition.offset);
+      }
+    } else {
+      newPosition = TextPosition(
+          offset: child.container.documentOffset + newPosition.offset);
+    }
+    return newPosition;
+  }
+
+  /// Returns the TextPosition below the given offset into the text.
+  ///
+  /// If the offset is already on the last line, the offset of the last
+  /// character will be returned.
+  @override
+  TextPosition getTextPositionBelow(TextPosition position) {
+    final child = childAtPosition(position);
+    final localPosition =
+        TextPosition(offset: position.offset - child.container.documentOffset);
+
+    var newPosition = child.getPositionBelow(localPosition);
+
+    if (newPosition == null) {
+      // There was no text above in the current child, check the direct
+      // sibling.
+      final sibling = childAfter(child);
+      if (sibling == null) {
+        // reached beginning of the document, move to the
+        // last character
+        newPosition = TextPosition(offset: document.length - 1);
+      } else {
+        final caretOffset = child.getOffsetForCaret(localPosition);
+        const testPosition = TextPosition(offset: 0);
+        final testOffset = sibling.getOffsetForCaret(testPosition);
+        final finalOffset = Offset(caretOffset.dx, testOffset.dy);
+        final siblingPosition = sibling.getPositionForOffset(finalOffset);
+        newPosition = TextPosition(
+            offset: sibling.container.documentOffset + siblingPosition.offset);
+      }
+    } else {
+      newPosition = TextPosition(
+          offset: child.container.documentOffset + newPosition.offset);
+    }
+    return newPosition;
+  }
+
+  // End TextLayoutMetrics implementation
+
+  QuillVerticalCaretMovementRun startVerticalCaretMovement(
+      TextPosition startPosition) {
+    return QuillVerticalCaretMovementRun._(
+      this,
+      startPosition,
+    );
+  }
+
+  @override
+  void systemFontsDidChange() {
+    super.systemFontsDidChange();
+    markNeedsLayout();
+  }
+}
+
+class QuillVerticalCaretMovementRun
+    extends BidirectionalIterator<TextPosition> {
+  QuillVerticalCaretMovementRun._(
+    this._editor,
+    this._currentTextPosition,
+  );
+
+  TextPosition _currentTextPosition;
+
+  final RenderEditor _editor;
+
+  @override
+  TextPosition get current {
+    return _currentTextPosition;
+  }
+
+  @override
+  bool moveNext() {
+    _currentTextPosition = _editor.getTextPositionBelow(_currentTextPosition);
+    return true;
+  }
+
+  @override
+  bool movePrevious() {
+    _currentTextPosition = _editor.getTextPositionAbove(_currentTextPosition);
+    return true;
+  }
 }
 
 class EditableContainerParentData
     extends ContainerBoxParentData<RenderEditableBox> {}
 
+/// Multi-child render box of editable content.
+///
+/// Common ancestor for [RenderEditor] and [RenderEditableTextBlock].
 class RenderEditableContainerBox extends RenderBox
     with
         ContainerRenderObjectMixin<RenderEditableBox,
@@ -1132,12 +1647,14 @@ class RenderEditableContainerBox extends RenderBox
         RenderBoxContainerDefaultsMixin<RenderEditableBox,
             EditableContainerParentData> {
   RenderEditableContainerBox(
-    List<RenderEditableBox>? children,
-    this._container,
-    this.textDirection,
-    this.scrollBottomInset,
-    this._padding,
-  ) : assert(_padding.isNonNegative) {
+      {required container_node.Container container,
+      required this.textDirection,
+      required this.scrollBottomInset,
+      required EdgeInsetsGeometry padding,
+      List<RenderEditableBox>? children})
+      : assert(padding.isNonNegative),
+        _container = container,
+        _padding = padding {
     addAll(children);
   }
 
@@ -1147,9 +1664,7 @@ class RenderEditableContainerBox extends RenderBox
   double scrollBottomInset;
   EdgeInsets? _resolvedPadding;
 
-  container_node.Container getContainer() {
-    return _container;
-  }
+  container_node.Container get container => _container;
 
   void setContainer(container_node.Container c) {
     if (_container == c) {
@@ -1172,7 +1687,7 @@ class RenderEditableContainerBox extends RenderBox
 
   EdgeInsets? get resolvedPadding => _resolvedPadding;
 
-  void _resolvePadding() {
+  void resolvePadding() {
     if (_resolvedPadding != null) {
       return;
     }
@@ -1185,11 +1700,11 @@ class RenderEditableContainerBox extends RenderBox
   RenderEditableBox childAtPosition(TextPosition position) {
     assert(firstChild != null);
 
-    final targetNode = _container.queryChild(position.offset, false).node;
+    final targetNode = container.queryChild(position.offset, false).node;
 
     var targetChild = firstChild;
     while (targetChild != null) {
-      if (targetChild.getContainer() == targetNode) {
+      if (targetChild.container == targetNode) {
         break;
       }
       final newChild = childAfter(targetChild);
@@ -1209,15 +1724,20 @@ class RenderEditableContainerBox extends RenderBox
     markNeedsLayout();
   }
 
-  RenderEditableBox? childAtOffset(Offset offset) {
+  /// Returns child of this container located at the specified local `offset`.
+  ///
+  /// If `offset` is above this container (offset.dy is negative) returns
+  /// the first child. Likewise, if `offset` is below this container then
+  /// returns the last child.
+  RenderEditableBox childAtOffset(Offset offset) {
     assert(firstChild != null);
-    _resolvePadding();
+    resolvePadding();
 
     if (offset.dy <= _resolvedPadding!.top) {
-      return firstChild;
+      return firstChild!;
     }
     if (offset.dy >= size.height - _resolvedPadding!.bottom) {
-      return lastChild;
+      return lastChild!;
     }
 
     var child = firstChild;
@@ -1230,7 +1750,7 @@ class RenderEditableContainerBox extends RenderBox
       dy += child.size.height;
       child = childAfter(child);
     }
-    throw 'No child';
+    throw StateError('No child at offset $offset.');
   }
 
   @override
@@ -1245,7 +1765,7 @@ class RenderEditableContainerBox extends RenderBox
   @override
   void performLayout() {
     assert(constraints.hasBoundedWidth);
-    _resolvePadding();
+    resolvePadding();
     assert(_resolvedPadding != null);
 
     var mainAxisExtent = _resolvedPadding!.top;
@@ -1291,7 +1811,7 @@ class RenderEditableContainerBox extends RenderBox
 
   @override
   double computeMinIntrinsicWidth(double height) {
-    _resolvePadding();
+    resolvePadding();
     return _getIntrinsicCrossAxis((child) {
       final childHeight = math.max<double>(
           0, height - _resolvedPadding!.top + _resolvedPadding!.bottom);
@@ -1303,7 +1823,7 @@ class RenderEditableContainerBox extends RenderBox
 
   @override
   double computeMaxIntrinsicWidth(double height) {
-    _resolvePadding();
+    resolvePadding();
     return _getIntrinsicCrossAxis((child) {
       final childHeight = math.max<double>(
           0, height - _resolvedPadding!.top + _resolvedPadding!.bottom);
@@ -1315,7 +1835,7 @@ class RenderEditableContainerBox extends RenderBox
 
   @override
   double computeMinIntrinsicHeight(double width) {
-    _resolvePadding();
+    resolvePadding();
     return _getIntrinsicMainAxis((child) {
       final childWidth = math.max<double>(
           0, width - _resolvedPadding!.left + _resolvedPadding!.right);
@@ -1327,7 +1847,7 @@ class RenderEditableContainerBox extends RenderBox
 
   @override
   double computeMaxIntrinsicHeight(double width) {
-    _resolvePadding();
+    resolvePadding();
     return _getIntrinsicMainAxis((child) {
       final childWidth = math.max<double>(
           0, width - _resolvedPadding!.left + _resolvedPadding!.right);
@@ -1339,7 +1859,7 @@ class RenderEditableContainerBox extends RenderBox
 
   @override
   double computeDistanceToActualBaseline(TextBaseline baseline) {
-    _resolvePadding();
+    resolvePadding();
     return defaultComputeDistanceToFirstActualBaseline(baseline)! +
         _resolvedPadding!.top;
   }
