@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 
 import '../../../flutter_quill.dart';
-import 'quill_icon_button.dart';
 
 class IndentButton extends StatefulWidget {
   const IndentButton({
@@ -39,6 +38,7 @@ class _IndentButtonState extends State<IndentButton> {
       size: widget.iconSize * 1.77,
       icon: Icon(widget.icon, size: widget.iconSize, color: iconColor),
       fillColor: iconFillColor,
+      borderRadius: widget.iconTheme?.borderRadius ?? 2,
       onPressed: () {
         final indent = widget.controller
             .getSelectionStyle()
