@@ -5,7 +5,6 @@ import '../../models/documents/style.dart';
 import '../../models/themes/quill_icon_theme.dart';
 import '../controller.dart';
 import '../toolbar.dart';
-import 'quill_icon_button.dart';
 
 typedef ToggleStyleButtonBuilder = Widget Function(
   BuildContext context,
@@ -146,5 +145,6 @@ Widget defaultToggleStyleButtonBuilder(
     icon: Icon(icon, size: iconSize, color: iconColor),
     fillColor: fill,
     onPressed: onPressed,
+    borderRadius: iconTheme?.borderRadius ?? 2,
   );
 }
