@@ -1,15 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
 
-import '../../models/documents/nodes/embed.dart';
+import '../../models/documents/nodes/embeddable.dart';
 import '../../models/themes/quill_dialog_theme.dart';
 import '../../models/themes/quill_icon_theme.dart';
-import '../../utils/media_pick_setting.dart';
 import '../controller.dart';
-import '../link_dialog.dart';
 import '../toolbar.dart';
-import 'image_video_utils.dart';
-import 'quill_icon_button.dart';
 
 class VideoButton extends StatelessWidget {
   const VideoButton({
@@ -59,6 +55,7 @@ class VideoButton extends StatelessWidget {
       hoverElevation: 0,
       size: iconSize * 1.77,
       fillColor: iconFillColor,
+      borderRadius: iconTheme?.borderRadius ?? 2,
       onPressed: () => _onPressedHandler(context),
     );
   }
